fix: move network links file, add persons to contact page
