+++ src/components/illustrations/networks-links.tsx
@@ -1,8 +1,8 @@
 import React, { useMemo } from 'react';
 import { Box, Flex, Link } from 'theme-ui';
-import { SocialFacebook } from '../../illustrations/social-facebook';
-import { SocialGithub } from '../../illustrations/social-github';
-import { SocialLinkedin } from '../../illustrations/social-linkedin';
+import { SocialFacebook } from './social-facebook';
+import { SocialGithub } from './social-github';
+import { SocialLinkedin } from './social-linkedin';
 
 export interface NetworksLinksProps {
   githubUrl?: string;
