+++ src/pages/contact.tsx
@@ -1,5 +1,5 @@
-import { Grid, Text } from '@theme-ui/components';
-import { Person } from '../components/contact/Person';
+import { Grid, Text } from 'theme-ui';
+import { Person } from '../components/contact/person';
 import { Layout } from '../components/layout/layout';
 
 const Contact: React.FC = () => {
@@ -13,12 +13,25 @@ const Contact: React.FC = () => {
           github={'https://github.com/braaar'}
           linkedIn={'https://www.linkedin.com/in/braaar/'}
         />
-        <Text>Simen</Text>
-        <Text>Brage</Text>
-        <Text>BNS</Text>
+        <Person
+          name={'Simen A. W. Olsen'}
+          photoPath={'/persons/simen-olsen.png'}
+          email={'so@bjerk.io'}
+          github={'https://github.com/cobraz'}
+          linkedIn={'https://www.linkedin.com/in/cobraz/'}
+        />
+        <Person
+          name={'Bjørn Niklas Sjøstrøm'}
+          photoPath={'/persons/bjorn-niklas-sjostrom.jpg'}
+          email={'bns@bjerk.io'}
+          linkedIn={
+            'https://www.linkedin.com/in/bj%C3%B8rn-niklas-sj%C3%B8str%C3%B8m-8043277/'
+          }
+        />
       </Grid>
     </Layout>
   );
 };
 
+// eslint-disable-next-line import/no-default-export
 export default Contact;
